Rename booster-task-profile to booster-task-analyser
