+++ booster-task-analyser/src/main/kotlin/com/didiglobal/booster/task/analyser/graph/CallGraphFormatter.kt
@@ -1,0 +1,12 @@
+package com.didiglobal.booster.task.analyser.graph
+
+/**
+ * Represents a call graph formatter
+ *
+ * @author johnsonlee
+ */
+interface CallGraphFormatter {
+
+    fun format(graph: CallGraph): CharSequence
+
+}
