+++ booster-task-analyser/src/main/resources/whitelist.txt
