+++ booster-task-analyser/src/test/java/org/greenrobot/eventbus/Subscribe.java
