+++ booster-task-analyser/src/test/kotlin/com/didiglobal/booster/task/analyser/AnnotationExample1.kt
@@ -1,0 +1,20 @@
+package com.didiglobal.booster.task.analyser
+
+import androidx.annotation.MainThread
+
+@MainThread
+class AnnotationExample1 {
+
+    fun fun1() {
+    }
+
+    fun fun2() {
+    }
+
+    fun fun3() {
+    }
+
+    fun fun4() {
+    }
+
+}
