+++ booster-task-analyser/src/test/resources/AnnotationExample1.class
