+++ booster-task-analyser/src/test/resources/AnnotationExample2.class
