--- booster-task-profile/README.md
@@ -1,56 +1,0 @@
-# booster-task-profile
-
-This module is used for potential performance issues detecting.
-
-## Getting Started
-
-Start profiling by executing the following command
-
-```bash
-./gradlew profileDebug
-```
-
-or 
-
-```bash
-./gradlew profileRelease
-```
-
-or
-
-```bash
-./gradlew profile
-```
-
-## Profile Reports
-
-The [dot](https://www.graphviz.org/doc/info/lang.html) format reports is located at `build/reports/${variant}/booster-task-profile/`,  you can convert the dot files to PNGs by using the following command:
-
-```bash
-find build/reports -name '*.dot' | xargs -t -I{} dot -O -Tpng {}
-```
-
-Here is an example generated by dot:
-
-![com.didiglobal.booster.demo.MainActivity](../assets/com.didiglobal.booster.demo.MainActivity.dot.png)
-
-## Properties
-
-The following table shows the properties that transformer supports:
-
-| Property                         | Description                                                                                   | Example                             |
-| -------------------------------- | --------------------------------------------------------------------------------------------- | ----------------------------------- |
-| `booster.task.profile.blacklist` | URI of API black list（Using [built-in API list](src/main/resources/blacklist.txt) by default）| file:///Users/booster/blacklist.txt |
-| `booster.task.profile.whitelist` | URI of API white list（Using [built-in API list](src/main/resources/whitelist.txt) by default）| file:///Users/booster/whitelist.txt |
-
-The properties can be passthrough the command line as following:
-
-```bash
-./gradlew assembleDebug -Pbooster.task.profile.whitelist=file:///Users/booster/whitelist.txt
-```
-
-or configured in the `gradle.properties`:
-
-```properties
-booster.task.profile.whitelist=file:///Users/booster/whitelist.txt
-```
