--- booster-task-profile/src/main/kotlin/com/didiglobal/booster/task/profile/LayoutHandler.kt
@@ -1,17 +1,0 @@
-package com.didiglobal.booster.task.profile
-
-import org.xml.sax.Attributes
-import org.xml.sax.helpers.DefaultHandler
-
-/**
- * @author johnsonlee
- */
-class LayoutHandler : DefaultHandler() {
-
-    val views = mutableSetOf<String>()
-
-    override fun startElement(uri: String, localName: String, qName: String, attributes: Attributes) {
-        views += qName
-    }
-
-}
