--- booster-task-profile/src/main/kotlin/com/didiglobal/booster/task/profile/cha/AbstractClassSet.kt
@@ -1,19 +1,0 @@
-package com.didiglobal.booster.task.profile.cha
-
-import com.didiglobal.booster.transform.asm.className
-import org.objectweb.asm.tree.ClassNode
-
-/**
- * @author johnsonlee
- */
-abstract class AbstractClassSet : ClassSet {
-
-    override fun contains(element: ClassNode) = contains(element.className)
-
-    override fun containsAll(elements: Collection<ClassNode>) = elements.all {
-        contains(it.className)
-    }
-
-    override fun load() = this
-
-}
