--- booster-task-profile/src/main/kotlin/com/didiglobal/booster/task/profile/graph/CallGraphFormatter.kt
@@ -1,12 +1,0 @@
-package com.didiglobal.booster.task.profile.graph
-
-/**
- * Represents a call graph formatter
- *
- * @author johnsonlee
- */
-interface CallGraphFormatter {
-
-    fun format(graph: CallGraph): CharSequence
-
-}
