--- booster-task-profile/src/test/kotlin/com/didiglobal/booster/task/profile/AnnotationExample1.kt
@@ -1,20 +1,0 @@
-package com.didiglobal.booster.task.profile
-
-import androidx.annotation.MainThread
-
-@MainThread
-class AnnotationExample1 {
-
-    fun fun1() {
-    }
-
-    fun fun2() {
-    }
-
-    fun fun3() {
-    }
-
-    fun fun4() {
-    }
-
-}
