--- booster-task-profile/src/test/kotlin/com/didiglobal/booster/task/profile/AnnotationExample2.kt
@@ -1,24 +1,0 @@
-package com.didiglobal.booster.task.profile
-
-import org.greenrobot.eventbus.Subscribe
-import org.greenrobot.eventbus.ThreadMode
-
-class AnnotationExample2 {
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun fun1() {
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun fun2() {
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun fun3() {
-    }
-
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    fun fun4() {
-    }
-
-}
