+++ settings.gradle
@@ -21,6 +21,7 @@ include ':booster-api'
 include ':booster-build'
 include ':booster-command'
 include ':booster-gradle-plugin'
+include ':booster-task-analyser'
 include ':booster-task-check-snapshot'
 include ':booster-task-compression'
 include ':booster-task-compression-cwebp'
@@ -29,7 +30,6 @@ include ':booster-task-compression-processed-res'
 include ':booster-task-list-artifact'
 include ':booster-task-list-permission'
 include ':booster-task-list-shared-library'
-include ':booster-task-profile'
 include ':booster-task-resource-deredundancy'
 include ':booster-task-spi'
 include ':booster-transform-activity-thread'
